feat(Public sync): Resolve public gateway address
This change is needed to allow users of the Android Gateway to view/change their public gateway by providing a unified address across all bindings (PoWeb, PoHTTP, CogRPC).

That is, we'll be able to use public addresses like `frankfurt.relaycorp.cloud` and each client will be able to resolve that to an actual URL that they can use. For example, the CogRPC record for `frankfurt.relaycorp.cloud` currently resolves to `cogrpc-frankfurt.relaycorp.cloud`.

We can't merge this until the respective change in the Android Gateway is available.