fix: Reflect WiFi hotspot availability on Android 13 (#573)
# What kind of change does this PR introduce?
Fixes the hotspot state detection on Android 13.

# What is the current behavior?
This is a temporary fix for [#569 ](https://github.com/relaycorp/relaynet-courier-android/issues/569) but a long-term solution is still needed.

# What is the new behavior (if this is a feature change)?
On Android 13, we are looking to the gateway addresses request results to infer if hotspot is available. 
In previous versions of Android, there is no behavior change.